Refactor to MailController
