+++ api/v1/controllers/RegistrationController.js
@@ -352,26 +352,6 @@ function filterAttendees(req, res, next) {
 		});
 }
 
-function sendMailinglist(req, res, next) {
-	var listName = req.body.listName;
-	var mailList = mail.lists[listName]
-	var template = req.body.template;
-	
-	services.MailService.sendToList(mailList, template)
-		.then(function () {
-			res.body = {};
-
-			next();
-			return null;
-		})
-		.catch(function (error) {
-			next(error);
-			return null;
-		});
-}
-
-router.use(bodyParser.json());
-router.use(middleware.auth);
 
 router.post('/mentor', middleware.request(requests.MentorRequest),
 	middleware.permission(roles.NONE, _isAuthenticated), createMentor);
@@ -395,9 +375,6 @@ router.put('/attendee/decision/:id',  middleware.request(requests.AttendeeDecisi
 	middleware.permission(roles.ORGANIZERS), updateAttendeeDecision);
 router.put('/attendee/:id(\\d+)', middleware.request(requests.AttendeeRequest),
 	middleware.permission(roles.ORGANIZERS), updateAttendeeById);
-
-router.put('/sendlist', middleware.request(requests.SendListRequest), 
-	middleware.permission(roles.ORGANIZERS), sendMailinglist);
 
 router.use(middleware.response);
 router.use(middleware.errors);
