+++ api/v1/index.js
@@ -29,6 +29,7 @@ v1.use('/ecosystem', controllers.EcosystemController.router);
 v1.use('/health', controllers.HealthController.router);
 v1.use('/rsvp', controllers.RSVPController.router);
 v1.use('/stats', controllers.StatsController.router);
+v1.use('/mail', controllers.MailController.router);
 
 // log any outgoing response for debugging
 v1.use(function (req, res, next) {
